requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.0
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.11.0
async-timeout==5.0.1
attrs==25.4.0
backports.asyncio.runner==1.2.0
certifi==2025.10.5
charset-normalizer==3.4.3
click==8.3.0
colorama==0.4.6
datasets==4.1.1
dill==0.4.0
distro==1.9.0
dotenv==0.9.9
exceptiongroup==1.3.0
filelock==3.19.1
frozenlist==1.8.0
fsspec==2025.9.0
griffe==1.14.0
h11==0.16.0
hf-xet==1.1.10
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.35.3
idna==3.10
iniconfig==2.1.0
jiter==0.11.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
mcp==1.16.0
multidict==6.7.0
multiprocess==0.70.16
numpy==2.2.6
openai==1.109.1
-e git+https://github.com/CH-chuan/openai-agents-self.git@6633fec6cad048832388a18554c9d2bd1926f105#egg=openai_agents
packaging==25.0
pandas==2.3.3
pluggy==1.6.0
propcache==0.4.0
pyarrow==21.0.0
pydantic==2.11.10
pydantic-settings==2.11.0
pydantic_core==2.33.2
Pygments==2.19.2
pytest==8.4.2
pytest-asyncio==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.3
referencing==0.36.2
requests==2.32.5
rpds-py==0.27.1
six==1.17.0
sniffio==1.3.1
sse-starlette==3.0.2
starlette==0.48.0
tomli==2.2.1
tqdm==4.67.1
types-requests==2.32.4.20250913
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.37.0
xxhash==3.6.0
yarl==1.22.0
